Ai/final/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.2
python-dotenv==1.2.2
numpy==2.4.4
faiss-cpu==1.13.2
sentence-transformers==5.4.1
langchain==1.2.15
langchain-core==1.3.0
langchain-google-genai==4.2.2